% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avelplcf.R
\name{avelplcf}
\alias{avelplcf}
\title{An ``average velocity'' of a \code{plcf} object.}
\usage{
avelplcf(obj, lag)
}
\arguments{
\item{obj}{A \code{plcf} object}

\item{The}{lag to use in the average velocity. See above formulas.}
}
\value{
\code{avelplcf} returns a \code{plcf} object. If the input \code{obj}
was defined from t_0 to t_f then the output will be defined from t_0+lag/2 to 
t_f+lag/2.
}
\description{
If f(t) represents the function underlying a \code{plcf} object, then this is 
[f(t+lag/2)-f(t-lag/2)]/lag, which is another \code{plcf}.
}
\examples{
#need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
