% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corlcfs.R
\name{corlcfs}
\alias{corlcfs}
\title{Correlation of two \code{lcf} object}
\usage{
corlcfs(obj1, obj2, bds = NULL)
}
\arguments{
\item{obj1}{An \code{lcf} object}

\item{obj2}{Another such}

\item{bds}{Bounds over which to compute the correlation analogue. Default NULL 
means to use the intersection of the bounds over which each \code{lcf} is defined. 
If a non-NULL value is passed, each \code{lcf} object must be defined over the
given range.}
}
\value{
\code{corlcfs} returns a single number which is the correlation.
}
\description{
Computes a continuous analogue of Pearson's correlation for two \code{lcf} object
}
\examples{
#need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
