% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplcfs.R
\name{corplcfs}
\alias{corplcfs}
\title{Correlation of two \code{plcf} object}
\usage{
corplcfs(obj1, obj2, bds = NULL)
}
\arguments{
\item{obj1}{A \code{plcf} object}

\item{obj2}{Another such}

\item{bds}{Bounds over which to compute the correlation analogue. Default NULL 
means to use the intersection of the bounds over which each \code{plcf} is defined. 
If a non-NULL value is passed, each \code{plcf} object must be defined over the
given range.}
}
\value{
\code{corplcfs} returns a single number which is the correlation.
}
\description{
Computes a continuous analogue of Pearson's correlation for two \code{plcf} objects
}
\examples{
#need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
