% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivplcf.R
\name{derivplcf}
\alias{derivplcf}
\alias{derivplcf.default}
\alias{derivplcf.plcf}
\title{Method for the \code{plcf} class that takes the derivative}
\usage{
derivplcf(object)

\method{derivplcf}{default}(object)

\method{derivplcf}{plcf}(object)
}
\arguments{
\item{object}{A \code{plcf} object}
}
\value{
\code{derivplcf} returns an object of class \code{lcf}.
}
\description{
The derivative of a \code{plcf} object is an \code{lcf} object.
}
\examples{
x<-c(1,2,4,12,13)
y<-c(0,1,-1,100,5)
obj<-plcf(x,y)
res<-derivplcf(obj)

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
