% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheckplcf.R
\name{errcheckplcf}
\alias{errcheckplcf}
\title{Looks to see if the \code{plcf} of a fish seems faulty}
\usage{
errcheckplcf(obj, reldatetime, upstreamthresh, timethresh)
}
\arguments{
\item{obj}{A \code{plcf} object}

\item{reldatetime}{Date and time of the release of the fish in the same numeric units
used in \code{obj$bpts}}

\item{upstreamthresh}{A threshold distance (km) of all-at-once upstream motion that 
disqualifies a record as unrealistic}

\item{timethresh}{A threshold time (days) without being detected that is tolerated 
before a record is recommended to remove}
}
\value{
\code{errcheckplcf} returns a numeric code indicating whether the record seems
faulty and if so what seems wrong with it. Codes are:
N: no problem detected
B: detections before the release
U: went upsteam, in one shot, more than \code{upstreamthresh}
T: went for too long without being detected
It is possible multiple problems occur; the error codes are concatenated.
}
\description{
Flags \code{plcf} records that seem faulty in various ways. Takes a single 
\code{plcf} and returns a code indicating whether it may be faulty and how.
}
\examples{
#Not written yet but need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
