% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalatpt.R
\name{evalatpt}
\alias{evalatpt}
\alias{evalatpt.plcf}
\alias{evalatpt.lcf}
\alias{evalatpt.default}
\title{Evaluates a \code{plcf} or \code{lcf} object at a point.
 
Evaluates a \code{plcf} or \code{lcf} object at a point.}
\usage{
evalatpt(object, ...)

\method{evalatpt}{plcf}(object, x)

\method{evalatpt}{lcf}(object, x)

\method{evalatpt}{default}(object, ...)
}
\arguments{
\item{x}{The point at which to evaluate it.}

\item{obj}{A \code{plcf} or \code{lcf} object.}
}
\value{
\code{evalatpt} returns the values of the function
represented by \code{object} at \code{x}. If \code{object}
is an \code{lcf} object and \code{x} equals one of the 
breakpoints, then NA is returned. If \code{x} is outside 
the range spanned by the breakpoints, NA is returned.
}
\description{
Evaluates a \code{plcf} or \code{lcf} object at a point.
 
Evaluates a \code{plcf} or \code{lcf} object at a point.
}
\examples{
#Not written yet but need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
