% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcf.R
\name{lcf}
\alias{lcf}
\title{Creator function for the \code{lcf} class}
\usage{
lcf(bpts, vals)
}
\arguments{
\item{bpts}{A numeric vector of the breakpoints between constant pieces. Must be in strictly ascending order.}

\item{vals}{A numeric vector with the values of the function between the breakpoints. Must be of length one less than \code{bpts}.}
}
\value{
\code{lcf} returns an object of class \code{lcf}. Slots are: 
\item{bpts}{The input} 
\item{vals}{The input}
}
\description{
The \code{lcf} class represents locally constant functions.
\code{lcf} inherits from the \code{list} class.
}
\examples{
x<-1:5
y<-c(.1,5,-3,4)
res<-lcf(x,y)

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
