% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plcf.R
\name{plcf}
\alias{plcf}
\title{Creator function for the \code{plcf} class}
\usage{
plcf(bpts, bptvals)
}
\arguments{
\item{bpts}{A numeric vector of the breakpoints between linear pieces}

\item{bptvals}{A numeric vector with the values of the function at the breakpoints}
}
\value{
\code{plcf} returns an object of class \code{plcf}. Slots are: 
\item{bpts}{The input, sorted into ascending order} 
\item{bptvals}{The input, rearranged so it still corresponds to the elements of \code{bpts}}
}
\description{
The \code{plcf} class represents piecewise linear continuous functions.
\code{plcf} inherits from the \code{list} class.
}
\examples{
x<-1:5
y<-c(.1,5,-3,4,10)
res<-plcf(x,y)

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
