% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotplcfs.R
\name{plotplcfs}
\alias{plotplcfs}
\title{Plots multiple \code{plcf} objects on the same axes}
\usage{
plotplcfs(plcflist, bds = NULL, xlabel, ylabel, filename = NULL)
}
\arguments{
\item{plcflist}{A list of \code{plcf} objects}

\item{bds}{Plot bounds for the x axis. Default value NULL corresponds to using the minimum range that includes the ranges of all the \code{plcf} objects.}

\item{filename}{File name (without extension) for saving the plot as a pdf. Default NULL uses the default plotting device instead.}
}
\value{
\code{plotplcfs} returns nothing but generates a plot.
}
\description{
Plots multiple \code{plcf} objects on the same axes
}
\examples{

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
