% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlcfs.R
\name{plotlcfs}
\alias{plotlcfs}
\title{Plots multiple \code{lcf} objects on the same axes}
\usage{
plotlcfs(lcflist, bds, filename)
}
\arguments{
\item{lcflist}{A list of \code{lcf} objects}

\item{bds}{Plot bounds for the x axis. Default value NA corresponds to using the minimum range that includes the ranges of all the \code{lcf} objects.}

\item{filename}{File name (without extension) for saving the plot as a pdf. Default NA uses the default plotting device instead.}
}
\value{
\code{plotlcfs} returns nothing but generates a plot.
}
\description{
Plots multiple \code{lcf} objects on the same axes
}
\examples{

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
